text_analytics/requirements.txt
pytesseract
Pillow
bs4
tesseract
opencv-python
PyPDF2
textract
pdf2image
nltk
autocorrect
wordcloud